Remove all superfluous animation used on the site, and libraries to support these animations
Main motivation of this change is that the mootools library (heavily used to support the animations)
is conflicting with a Flash emulator (Ruffle) I'm hoping to use in order to play the two flash files on
my site (he's an example of the issue: https://github.com/ruffle-rs/ruffle/issues/8619).

Also, the animations don't bring that much awesomeness, and could be distracting to some.
So, no reason yet to find another way to implement these animations.