Start using the Ruffle Flash emulator on the two pages with a Flash app
which allows these Flash apps to perfectly play! (at least on both Chrome and Firefox on a mac, which is all I've tested so far).

Note that we're currently using the latest, cutting edge, build of Ruffle.  And we're getting it from a CDN.
I hope to address both of these details in a future commit in order to make this use of Ruffle future proof of external changes.