Leverage Docker for testing website in a dev environment
Because it turns out that Github Pages doesn't use either Gemfile
or Gemfile.lock in order to build the appropriate Ruby environment,
see https://stackoverflow.com/a/56297531.

That, and the fact that it's finicky to manage a local Ruby dev
environment for testing this site, drives us to use Docker for
dev testing.

Here we're adding a Dockerfile which defines a docker image that's
appropriate for dev/local testing a Jekyll site that'll be deployed to
Github Pages.

Also, the use of Docker for local testing means the existing
Gemfile/Gemfile.lock are no longer needed.