Make an attempt at being mobile friendly
in the layout of this site.

Make use of flex and grid layouts in order accomplish
the same layouts we used to have, but in a way that scales
comfortably with the rest of the site when seen on mobile...

... at least that's what I think I'm doing.  Things looks moderately
nice now when this site is seen in a 'mobile simulation' view via
desktop Chrome.  Once this is deployed I can easily test on a mobile
device to see how nice these changes actually are.