Generally add support for viewing this site on mobile devices
By making the entire site responsive to the viewport width
and supporting very small viewport widths like 320px.

Note: there are a few features of this site that still can't be interacted
with via touch: (1) WAD viewer, and (2) Samurai Flash game.
Attempts at either adding the support needed in these places or
documenting the lack of support will come in a future change.