No longer expose filenames in project URLs
We're currently depending upon the filenames of the project pages to
control the sorting of these projects (because Jekyll 3.* doesn't give
us a better tool to sort a custom collection). This means that the
project filenames may change whenever a new project is added, in order
to correctly place this new project in the right order with the rest of
the projects.

We don't want the URLs to these projects to change whenever a new
project is added though, because we want existing links to old projects
to continue to work.

So, we'll use the `permalink` setting for the 'projects' collection and
the `slug` setting for each project to give each project a permanent URL.

And our new use of `slug` means that we can now use `slug` wherever we
were previously using `project_id`.