Rename existing project files to make room for new
We'll be introducing a new project that we wish to appear at the top of
the project list. This means that this new project must have a filename
that sorts before all existing projects. To accomplish this we'll use
a prefix of '1-' for this new project and bump the number prefix on all
existing projects.

This commit just accomplishes the second part, bumping the number prefix
on all existing projects.